zuoye

+++ Spring-Data-Jpa-Test/src/main/java/com/jpa/dao/CustomerDao.java
@@ -11,5 +11,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
  * @Date 2019/11/11 15:57
  * @Version 1.0
  **/
-public interface CustomerDao extends JpaRepository<Customer,Long>, JpaSpecificationExecutor<Customer> {
+public interface CustomerDao extends JpaRepository<Customer,Integer>, JpaSpecificationExecutor<Customer> {
 }
